feat(pp): add preprocessor interface and C directive recognition
IPreprocessor      pure virtual, one method: process(tokens) -> tokens
PreprocessorBase   identity pass, for languages with no preprocessor
CPreprocessor      the C implementation
No inheritance relationship with the lexer. CLexer is chunk-parallel
and reads bytes. CPreprocessor is sequential and reads only the token
vector. The single point of coupling is the source view, which
CPreprocessor holds to recover spelling through token ranges.

+++ include/Z/Zaban/Lex/Token.hpp
@@ -24,7 +24,7 @@ namespace Z::Zaban::Lex {
         /// The source range occupied by the token.
         OffsetRange<OffsetType> range;
 
-        std::uint8_t flags = 0;
+        std::uint16_t flags = 0;
 
        public:
         /** @brief Constructs a token from its beginning and ending locations.
